A few conversions from float to int and defining constants as float (rather than double) to reduce warnings with Intel comiler.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@2129 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/layers/Image.hpp
@@ -49,7 +49,7 @@ public:
    static int convertToGrayScale(PVLayerLoc * loc, unsigned char * buf);
 
    int  convolve(int width);
-   void setTau(int t)                { tau = t; }
+   void setTau(float t)                { tau = t; }
 
    int copyFromInteriorBuffer(const unsigned char * buf);
    int copyToInteriorBuffer(unsigned char * buf);
